Seperated command utils helper functions
